refactor(metadata): group metadata as one class passed to recipes and operators
This cleans up the codebase a bit and means the JSON schema and OpenAPI models don't overlap and
rename themselves through pydantic.
